+++ src/setup.cpp
@@ -433,10 +433,8 @@ bool Setup::existsProgramsHome() {
   return isDir(LODA_HOME + "programs" + FILE_SEP + "oeis");
 }
 
-bool Setup::cloneProgramsHome(std::string git_url) {
-  const std::string git_clone =
-      "git clone " + git_url + " \"" + getLodaHome() + "programs\"";
-  return system(git_clone.c_str()) == 0;
+void Setup::cloneProgramsHome(std::string git_url) {
+  git("clone " + git_url + " \"" + getLodaHome() + "programs\"");
 }
 
 bool Setup::checkProgramsHome() {
@@ -444,25 +442,11 @@ bool Setup::checkProgramsHome() {
   if (!existsProgramsHome()) {
     std::cout << "LODA needs to download its programs repository from GitHub."
               << std::endl;
-    std::cout << "It contains programs for more than 50,000 integer sequences."
-              << std::endl;
-    std::cout << "It is needed to evaluate known integer sequence programs and "
-              << std::endl;
-    std::cout << "for mining new programs using the 'loda mine' command."
-              << std::endl;
     std::cout << "The repository requires around 350 MB of disk space."
-              << std::endl
               << std::endl;
-    if (!hasGit()) {
-      std::cout << std::endl
-                << "The setup requires the git tool to download the programs."
-                << std::endl;
-      std::cout
-          << "Please install git and restart the setup:" << std::endl
-          << "https://git-scm.com/book/en/v2/Getting-Started-Installing-Git"
-          << std::endl;
-      return false;
-    }
+    std::cout << "Checking whether git is installed:" << std::endl;
+    git("--version");
+    std::cout << std::endl;
     std::string git_url = "https://github.com/loda-lang/loda-programs.git";
     std::cout << "Press return to download the default programs repository:"
               << std::endl;
@@ -471,11 +455,7 @@ bool Setup::checkProgramsHome() {
     if (!line.empty()) {
       git_url = line;
     }
-    if (cloneProgramsHome(git_url)) {
-      std::cout << std::endl
-                << "Error cloning repository. Aborting setup." << std::endl;
-      return false;
-    }
+    cloneProgramsHome(git_url);
     std::cout << std::endl;
   }
   return true;